Add files via upload
The accuracy of our models is pretty bad. I found that using the means is probably not the best representation of our data as the model does not know it is a calculated mean and so it assumes one value for each feature. The main things we need to do are get our scores to look you know reasonable and Pearsons Correlation Coefficient for each features significance in the model.